refactor: improve some code
