+++ performance/src/components/header.tsx
@@ -1,4 +1,4 @@
-import { HeaderProps } from '$/types/types';
+import { HeaderProps } from '$/types';
 
 export function Header({ title_text }: HeaderProps) {
   return (
