+++ performance/src/router/index.ts
@@ -1,0 +1,1 @@
+export { Router } from './router.tsx';
